requirements.txt
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
PyQt5==5.15.4
scikit_learn==0.24.2
matplotlib==3.4.0